refactor: optimized the Xline startup process
