+++ xline/src/lib.rs
@@ -177,3 +177,5 @@ pub mod server;
 mod state;
 /// Storage module
 pub mod storage;
+/// Xline utils
+pub mod xline_utils;
